requirements.txt
boto3
fastapi
uvicorn
pydantic
python-docx
pyyaml
python-dotenv
streamlit
requests
faiss-cpu
tqdm
pdfminer.six
python-multipart
chromadb
langchain
langchain-community
langchain-openai
PyPDF2
opensearch-py